Clone helper - WIP.

+++ crafting_combinator_xeraph/control.lua
@@ -8,6 +8,7 @@ local util = require 'script.util'
 local gui = require 'script.gui'
 local blueprint = require 'script.blueprint'
 local migration_helper = require 'script.migration-helper'
+local clone_helper = require 'script.clone-helper'
 
 local cc_rate, rc_rate = 1, 1
 
@@ -22,10 +23,11 @@ end
 
 local function on_load(forced)
 	if not forced and next(late_migrations.__migrations) ~= nil then return; end
-	
+
 	cc_control.on_load()
 	rc_control.on_load()
 	signals.on_load()
+	clone_helper.on_load()
 	cc_rate = settings.global[config.REFRESH_RATE_CC_NAME].value
 	rc_rate = settings.global[config.REFRESH_RATE_RC_NAME].value
 	
@@ -43,13 +45,14 @@ end
 local function init_global()
 	global.delayed_blueprint_tag_state = global.delayed_blueprint_tag_state or {}
 	global.dead_combinator_settings = global.dead_combinator_settings or {}
+	global.clone_placeholder = global.clone_placeholder or {}
 end
 
 script.on_init(function()
+	init_global()
 	cc_control.init_global()
 	rc_control.init_global()
 	signals.init_global()
-	init_global()
 	on_load(true)
 end)
 script.on_load(on_load)
@@ -85,6 +88,25 @@ local function on_built(event)
 	blueprint.handle_event(event)
 end
 
+local function on_cloned(event)
+	---event.source
+	---event.destination
+
+	if entity.name == config.CC_NAME then
+	elseif entity.name == config.MODULE_CHEST_NAME then
+	elseif entity.name == config.RC_NAME then
+	elseif entity.name == config.RC_PROXY_NAME then
+	end
+
+	-- assembler and containers
+	local entity_type = entity.type
+	if entity_type == 'assembling-machine' then
+		cc_control.update_assemblers(entity.surface, entity);
+	elseif util.CONTAINER_TYPES[entity.type] then
+		cc_control.update_chests(entity.surface, entity);
+	end
+end
+
 local function save_dead_combinator_settings(uid, settings)
 	-- entry is created during on_entity_died and removed during on_post_entity_died
 	global.dead_combinator_settings[uid] = settings
@@ -105,23 +127,22 @@ local function on_destroyed(event) -- on_entity_died, on_player_mined_entity, on
 		cc_control.update_chests(entity_surface, entity, true)
 	end
 
-	-- Skip cc destroy - by on_entity_died or self-raised script_raised_destroy
-	if event.skip_cc_destroy then return end
-
 	if entity_name == config.CC_NAME then
 		local uid = entity.unit_number
 		local combinator = global.cc.data[uid]
 		if event_name == defines.events.on_entity_died then
 			save_dead_combinator_settings(uid, combinator.settings)
-			script.raise_event(defines.events.script_raised_destroy, {entity = combinator.module_chest, skip_cc_destroy = true})
-			combinator.module_chest.destroy()
 		elseif event_name == defines.events.on_player_mined_entity then
 			-- Need to mine module chest first, success == true
-			if not cc_control.mine_module_chest(entity.unit_number, event.player_index) then
+			if not cc_control.mine_module_chest(uid, event.player_index) then
 				-- Unable to mine module chest, cc cloned and replaced - remove cc from buffer, then return
 				event.buffer.remove({name=config.CC_NAME, count=1})
 				return
 			end
+		end
+		if event_name == defines.on_entity_died or event_name == defines.events.script_raised_destroy then
+			script.raise_event(defines.events.script_raised_destroy, {entity = combinator.module_chest, skip_cc_destroy = true})
+			combinator.module_chest.destroy()
 		end
 		cc_control.destroy(entity)
 	elseif entity_name == config.MODULE_CHEST_NAME then
@@ -131,11 +152,13 @@ local function on_destroyed(event) -- on_entity_died, on_player_mined_entity, on
 			event.buffer.remove({name=config.CC_NAME, count=1})
 		elseif event_name == defines.events.on_robot_mined_entity
 		or event_name == defines.events.script_raised_destroy then
-			-- Script_raised_destroy or mined by robot
 			-- This signifies that the module chest will be destroyed
-			-- Get cc_entity and raise script destroy for cc
-			local cc_entity = entity_surface.find_entity(config.CC_NAME, entity.position)
-			if cc_entity and cc_entity.valid then cc_entity.destroy({raise_destroy = true}) end
+			-- Skip cc destroy - by on_entity_died or self-raised script_raised_destroy
+			if not event.skip_cc_destroy then
+				-- Get cc_entity and raise script destroy for cc
+				local cc_entity = entity_surface.find_entity(config.CC_NAME, entity.position)
+				if cc_entity and cc_entity.valid then cc_entity.destroy({raise_destroy = true}) end
+			end
 		end
 	elseif entity_name == config.RC_NAME then
 		if event_name == defines.events.on_entity_died then
@@ -168,6 +191,11 @@ script.on_event(defines.events.on_tick, function(event)
 	
 	run_update(global.cc.ordered, event.tick, cc_rate)
 	run_update(global.rc.ordered, event.tick, rc_rate)
+end)
+
+script.on_nth_tick(600, function(event)
+	-- clean up partially cloned entities?
+	clone_helper.on_nth_tick(event)
 end)
 
 script.on_event(defines.events.on_player_rotated_entity, function(event)
@@ -214,6 +242,7 @@ script.on_event(defines.events.on_built_entity, on_built, filter_built_destroyed
 script.on_event(defines.events.on_robot_built_entity, on_built, filter_built_destroyed)
 script.on_event(defines.events.script_raised_built, on_built, filter_built_destroyed)
 script.on_event(defines.events.script_raised_revive, on_built, filter_built_destroyed)
+script.on_event(defines.events.on_entity_cloned, on_cloned, filter_built_destroyed)
 
 -- entity destroyed events
 script.on_event(defines.events.on_entity_died, on_destroyed, filter_built_destroyed)
